% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math_functions.R
\name{add}
\alias{add}
\alias{subtract}
\alias{multiply}
\alias{divide}
\title{Basic mathematical functions}
\usage{
add(x, y)

subtract(x, y)

multiply(x, y)

divide(x, y)
}
\arguments{
\item{x}{Numeric value}

\item{y}{Numeric value}
}
\value{
Numeric of length 1
}
\description{
Basic mathematical functions
}
\author{
Shona Wilde
}
