% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_group_identifier.R
\name{add_group_identifier}
\alias{add_group_identifier}
\title{Function to give unique identifiers to groups within data}
\usage{
add_group_identifier(df, type = "numbers", as_factor = F)
}
\arguments{
\item{df}{Input data frame or tibble}

\item{type}{Type of identifier. Either 'numbers' or 'letters'}

\item{as_factor}{Should the new column be a factor?}
}
\value{
Tibble
}
\description{
Function to give unique identifiers to groups within data
}
\author{
Shona Wilde
}
