% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_group_index.R
\name{add_group_index}
\alias{add_group_index}
\title{Function to add a group index to data}
\usage{
add_group_index(x, each)
}
\arguments{
\item{x}{Vector.}

\item{each}{Size of groups.}
}
\value{
Value.
}
\description{
\code{add_group_index} creates groups within data based on a repeating index.
}
\author{
Shona Wilde
}
