% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_nearest_date.R
\name{add_nearest_date}
\alias{add_nearest_date}
\title{Function to find and add nearest date in another data frame}
\usage{
add_nearest_date(df1, df2)
}
\arguments{
\item{df1}{Data frame.}

\item{df2}{Data frame to find nearest date in.}
}
\value{
Date vector of class \code{POSIXct}.
}
\description{
Function to find and add nearest date in another data frame
}
\author{
Shona Wilde
}
