% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_quantile.R
\name{add_quantile}
\alias{add_quantile}
\title{Add quantile}
\usage{
add_quantile(df, variable = "value", percent = 0.01, na.rm = FALSE)
}
\arguments{
\item{df}{Data frame}

\item{variable}{Variable to calculate quantile for}

\item{percent}{Quantile to calculate}

\item{na.rm}{Should NAs be removed?}
}
\value{
Tibble
}
\description{
Calculate single quantile for a dataset. Useful if wanting to filter below a certain threshold of values, ie lowest 1%.
}
\author{
Shona Wilde & Stuart Grange
}
