% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_logical_factor.R
\name{as_logical_factor}
\alias{as_logical_factor}
\title{Create a logical as an ordered factor}
\usage{
as_logical_factor(x, ...)
}
\arguments{
\item{x}{Object to convert}

\item{...}{Further arguments passed to or from other methods}
}
\value{
Ordered factor
}
\description{
Create an ordered factor with levels TRUE and FALSE
}
\author{
Shona Wilde
}
