% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign_time_of_day.R
\name{assign_time_of_day}
\alias{assign_time_of_day}
\title{Assign time of day to a date}
\usage{
assign_time_of_day(date)
}
\arguments{
\item{date}{POSIXct date}
}
\value{
Tibble
}
\description{
Function to assign a date as either 'morning', 'afternoon', 'evening' or 'night'
}
\author{
Shona Wilde
}
