% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_alpha.R
\name{calc_alpha}
\alias{calc_alpha}
\title{Calculate alpha}
\usage{
calc_alpha(heading, wd = "wd")
}
\arguments{
\item{heading}{Aircraft heading in degrees}

\item{wd}{Wind direction in degrees}
}
\value{
Angle in degrees.
}
\description{
Function to calculate the angle between wind direction and the line perpendicular to the flight track
}
\author{
Shona Wilde
}
