% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_flux.R
\name{calc_flux}
\alias{calc_flux}
\title{Calculate mass flux}
\usage{
calc_flux(
  wind_speed = ws,
  ground_speed = gspd_gin,
  alpha = alpha,
  scale_height = scale_height,
  mixing_height = mixing_height,
  plume_integral = plume_integral
)
}
\arguments{
\item{wind_speed}{Wind speed in ms-1}

\item{ground_speed}{Speed over ground in ms-1}

\item{alpha}{Angle between aircraft track and direction normal to wind direction}

\item{scale_height}{Scale height in m
Scale height is the increase in altitude for which the atmospheric pressure decreases by a factor of e}

\item{mixing_height}{Mixing height of observed plume in m}

\item{plume_integral}{Area under observed plume
This must be the enhancement above background concentrations}
}
\value{
value
}
\description{
Calculate emission rate of species using simplified mass flux technique in molecules per second
}
\author{
Shona Wilde
}
