% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_perpendicular_wind.R
\name{calc_perpendicular_wind}
\alias{calc_perpendicular_wind}
\title{Calculate perpendicular wind component}
\usage{
calc_perpendicular_wind(ws = "ws", alpha)
}
\arguments{
\item{ws}{Wind speed}

\item{alpha}{Angle between wind direction and the line perpendicular to the flight track 
calculated using \code{\link{calc_alpha}}.}
}
\value{
Numeric.
}
\description{
Function to calculate the component of horizontal wind speed perpendicular to the flight track
}
\author{
Shona Wilde
}
