% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_potential_temp.R
\name{calc_potential_temp}
\alias{calc_potential_temp}
\title{Calculate potential temperature}
\usage{
calc_potential_temp(temp, pressure, ref_pressure = 1000)
}
\arguments{
\item{temp}{Tempertaure in Kelvin}

\item{pressure}{Pressure in mb or hPa}

\item{ref_pressure}{Reference pressure to use in equation in mb or hPa}
}
\value{
value
}
\description{
Calculate potential temperature in Kelvin using Poissons Equation.
 \link{http://tornado.sfsu.edu/geosciences/classes/m201/PrimtiveEquations/DerivedRelationships.html}
}
\author{
Shona Wilde
}
