% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_uncertainty.R
\name{calc_relative_error}
\alias{calc_relative_error}
\title{Calculate uncertainty}
\usage{
calc_relative_error(x, blank_value)
}
\arguments{
\item{x}{Value for which uncertainty should be calculated}

\item{blank_value}{Concentration of variable in GC blanks in ppb}

\item{cal_rel_error}{Relative error from calibration standards calculated as the
calibration error in ppb divided by the concentration in ppb}

\item{detection_limit}{Limit of detection in ppb}

\item{work_std_precision}{Precision of woking standard calculated by dividing the standard deviation of the working standard 
by the mean concentration of the working standard in ppb}

\item{blank_error}{Relative error from blank values calculated using \code{calc_relative_error}}
}
\value{
value
}
\description{
Calculate uncertainty
}
\author{
Shona Wilde
}
