% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_scale_height.R
\name{calc_scale_height}
\alias{calc_scale_height}
\title{Calculate scale height}
\usage{
calc_scale_height(temperature)
}
\arguments{
\item{temperature}{Non de-iced temperature in Kelvin}
}
\value{
value
}
\description{
Calculate scale height for use in mass flux calculation. 
Scale height is the increase in altitude for which the atmospheric pressure decreases by a factor of e
}
\author{
Shona Wilde
}
