% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_wind.R
\name{calc_wind_speed}
\alias{calc_wind_speed}
\title{Calculate wind speed and direction}
\usage{
calc_wind_speed(u, v)
}
\arguments{
\item{u}{}

\item{v}{}
}
\value{
vector
}
\description{
Calculates wind speed and direction from u and v components of wind
}
\author{
Shona Wilde
}
