% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_acceleration.R
\name{calculate_acceleration}
\alias{calculate_acceleration}
\title{Function to calculate acceleration from speed and date variables}
\usage{
calculate_acceleration(speed, date)
}
\arguments{
\item{speed}{Speed in \eqn{m~s^{-1}}.}

\item{date}{Date and time variable.}
}
\value{
Value
}
\description{
Function to calculate acceleration from speed and date variables
}
\author{
Shona Wilde
}
