% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_group_split.R
\name{calculate_group_split}
\alias{calculate_group_split}
\title{Function to calculate the percentage of observations in each group}
\usage{
calculate_group_split(df, ...)
}
\arguments{
\item{df}{Data frame}

\item{...}{Grouping variable(s)}
}
\value{
Summary tibble
}
\description{
Function to calculate the percentage of observations in each group
}
\author{
Shona Wilde
}
