% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_percentiles.R
\name{calculate_percentiles}
\alias{calculate_percentiles}
\title{Calculate percentiles}
\usage{
calculate_percentiles(x, probs = c(0.01, 0.1, 0.25, 0.5, 0.75, 0.95,
  0.99))
}
\arguments{
\item{x}{numeric vector}

\item{probs}{numeric vector of quantiles to be calculated}
}
\value{
Tibble
}
\description{
Function to calculate summary statistics
}
\author{
Shona Wilde
}
