% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_response_factor.R
\name{calculate_response_factor}
\alias{calculate_response_factor}
\title{Calculate reponse factor}
\usage{
calculate_response_factor(area, concentration)
}
\arguments{
\item{area}{Peak area}

\item{concentration}{concentration in standard}
}
\value{
Numeric
}
\description{
Calculate reponse factor
}
\author{
Shona Wilde
}
