% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/centre.R
\name{centre}
\alias{centre}
\title{Function to centre or standardise a vector}
\usage{
centre(x, standardise = F)
}
\arguments{
\item{x}{Numeric vector}

\item{standardise}{Should \code{x} be rescaled to have a mean of 0 and a standard deviation of 1? Also known as the z-score}
}
\value{
Vector of same length as \code{x}
}
\description{
Function to centre or standardise a vector
}
\author{
Shona Wilde
}
