% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_variable_names.R
\name{clean_variable_column}
\alias{clean_variable_column}
\title{Function to clean variable strings contained in a column}
\usage{
df %>% 
     mutate(variable = clean_variable_column(variable, join_by = c("variable" = "variable_lookup"))
}
\arguments{
\item{x}{vector of names to clean}

\item{lookup_table}{A look-up table to be used for cleaning names}

\item{join_by}{Joining key in look-up table}
}
\value{
Tibble
}
\description{
Function to clean variable strings contained in a column
}
\author{
Shona Wilde
}
