% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_variable_names.R
\name{clean_variable_names}
\alias{clean_variable_names}
\title{Function to clean variable names in data frame}
\usage{
clean_variable_names(lookup, join_by = c("variable" = "variable_lookup"))
}
\arguments{
\item{df}{data}

\item{lookup_table}{A look-up table to be used for cleaning names}

\item{join_by}{Joining key in look-up table}
}
\value{
Tibble
}
\description{
Function to clean variable names in data frame
}
\author{
Shona Wilde
}
