% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cols_from_formula.R
\name{cols_from_formula}
\alias{cols_from_formula}
\title{Extract column names from a formula}
\usage{
cols_from_formula(formula)
}
\arguments{
\item{formula}{Formula object}
}
\value{
Vector of names
}
\description{
Extract column names from a formula
}
\author{
Shona Wilde
}
