% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_gps.R
\name{conv_gps}
\alias{conv_gps}
\title{Convert GPS}
\usage{
conv_gps(coord)
}
\arguments{
\item{coord}{lat/long}
}
\value{
value
}
\description{
Convert GPS data from WACL van to decimal degrees
}
\author{
Shona Wilde
}
