% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_critical_r.R
\name{correlation_summary}
\alias{correlation_summary}
\title{Correlation summary}
\usage{
correlation_summary(
  df,
  x = "x",
  y = "y",
  method = "pearson",
  conf_level = 0.95
)
}
\arguments{
\item{df}{Data frame containing model variables}

\item{x}{x-component for model}

\item{y}{y-component for model}

\item{method}{a character string indicating which correlation coefficient is to be used for the test.
One of "pearson", "kendall", or "spearman".}

\item{conf_level}{confidence level for the returned confidence interval}
}
\value{
Tibble
}
\description{
Function to calculate correlation coefficients and tidy output
}
\author{
Shona Wilde
}
