% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/country_from_coords.R
\name{country_from_coords}
\alias{country_from_coords}
\title{Function to obtain country from pair of coordinates}
\usage{
country_from_coords(data)
}
\arguments{
\item{data}{Data with class \code{sp} or \code{sf}}
}
\value{
Either an `sp` or `sf` object
}
\description{
Function to obtain country from pair of coordinates
}
\author{
Shona Wilde
}
