% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_empty_tibble.R
\name{create_empty_tibble}
\alias{create_empty_tibble}
\title{Function to create empty tibble with 0 rows}
\usage{
create_empty_tibble(names)
}
\arguments{
\item{names}{Vector of names for tibble}
}
\value{
Empty tibble with column names
}
\description{
Function to create empty tibble with 0 rows
}
\author{
Shona Wilde
}
