% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_group_index.R
\name{create_group_index}
\alias{create_group_index}
\title{Function to create groups within data}
\usage{
create_group_index(df, group_size)
}
\arguments{
\item{df}{Input data frame}

\item{group_size}{Size of groups to create}
}
\value{
Tibble
}
\description{
\code{create_group_index} creates groups within data based on a repeating index
}
\author{
Shona Wilde
}
