% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_data_capture.R
\name{data_capture}
\alias{data_capture}
\title{Calculate Percentage Data Capture}
\usage{
data_capture(x)
}
\arguments{
\item{x}{variable}
}
\value{
value
}
\description{
Calculates percentage of data capture over a period of time
}
\author{
Shona Wilde
}
