% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_key_pointrange_h.R
\name{draw_key_pointrange_h}
\alias{draw_key_pointrange_h}
\title{Draw horizontal pointrange key glyph in a \code{ggplot} legend}
\usage{
draw_key_pointrange_h(data, params, size)
}
\arguments{
\item{data}{Data to be displayed in layer}

\item{params}{Additional parameters to the geom and stat}

\item{size}{Size}
}
\description{
Draw horizontal pointrange key glyph in a \code{ggplot} legend
}
\author{
Shona Wilde
}
