% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_by_date_range.R
\name{filter_by_date_range}
\alias{filter_by_date_range}
\title{Filter a data frame between two dates}
\usage{
filter_by_date_range(df, start_date, end_date, id = NA)
}
\arguments{
\item{df}{Input data frame or tibble}

\item{start_date}{Start date}

\item{end_date}{End date}

\item{id}{Name of ID column to add to data. Useful when using with \code{\link{purrr::map}}}
}
\value{
Tibble
}
\description{
Filter a data frame between two dates
}
\author{
Shona Wilde
}
