% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/breakpoints.R
\name{find_breakpoints}
\alias{find_breakpoints}
\title{Function to detect breakpoints}
\usage{
find_breakpoints(df, h = 0.15, breaks = NULL)
}
\arguments{
\item{df}{Data frame to detect breakpoints in}

\item{h}{minimal segment size either given as fraction relative to the sample 
size or as an integer giving the minimal number of observations in each segment.}

\item{breaks}{Maximum number of breaks to be calculated}

\item{x}{Object of class \code{"breakpoints"}}
}
\value{
An object of class \code{"breakpoints"}
}
\description{
\code{find_breakpoints} can be used to find breakpoints which can be extracted as a vector by \code{get_break_numbers}
}
\author{
Shona Wilde
}
