% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_current_weather.R
\name{get_current_weather}
\alias{get_current_weather}
\title{Download current weather data from OpenWeatherMap}
\usage{
get_current_weather(api_key, lat = 53.96, lon = -1.0873)
}
\arguments{
\item{api_key}{API key for OpenWeatherMap}

\item{lat}{Latitude of location to get data for}

\item{lon}{Longitude of location to get data for}
}
\value{
Tibble
}
\description{
Download data from the \href{https://openweathermap.org/}{OpenWeatherMap API}
}
\author{
Shona Wilde
}
