% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_first_last.R
\name{get_first_last}
\alias{get_first_last}
\title{Function to get first and last elements of a vector}
\usage{
get_first_last(x)
}
\arguments{
\item{x}{Vector}
}
\value{
Tibble
}
\description{
Function to get first and last elements of a vector
}
\author{
Shona Wilde
}
