% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_flight_events.R
\name{get_flight_events}
\alias{get_flight_events}
\title{Split flight into range events}
\usage{
get_flight_events(flight_data, flight_sum)
}
\arguments{
\item{flight_data}{merge file containing flight data. Date must be called "date" and of class \code{POSICXct}}

\item{flight_sum}{FAAM flight summary read using \code{\link{read_faam_flight_summary}}}
}
\value{
Tibble
}
\description{
Splits up flight data into straight level runs and profiles which are defined in flight summary
}
\author{
Will Drysdale / Shona Wilde
}
