% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpolate_time_series.R
\name{interpolate_time_series}
\alias{interpolate_time_series}
\title{Interpolate values within a time series}
\usage{
interpolate_time_series(df, by = "hour")
}
\arguments{
\item{df}{Input data frame or tibble}

\item{by}{Increment of the new time series. Options are 'year', 'month', 'day', 'hour', 'min', or 'sec'}
}
\value{
Tibble
}
\description{
This function is typically used to increase the resolution of time series data e.g hourly to minute. All numeric variables within the data will be interpolated. 
Character or factor variables are filled down.
}
\author{
Shona Wilde
}
