% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lagged_change.R
\name{lagged_change}
\alias{lagged_change}
\alias{cumulative_change}
\title{Calculate the change along a vector}
\usage{
lagged_change(x, type = "percent")

cumulative_change(x, type = "percent")
}
\arguments{
\item{x}{Numeric vector}

\item{type}{Type of change to calculate. Either "percent" or "absolute"}
}
\value{
Numeric vector of the same length as \code{x}
}
\description{
\code{lagged_change()} calculates the change between consecutive lagged values of \code{x}.
\code{cumulative_change()} calculates the change between the first value of \code{x} and each subsequent value of \code{x}.
}
\author{
Shona Wilde
}
