% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_as_st.R
\name{map_as_st}
\alias{map_as_st}
\title{Function to extract world map as an \code{sf} object}
\usage{
map_as_st()
}
\value{
object of  class \code{sf} object
}
\description{
Function to extract world map as an \code{sf} object
}
\author{
Shona Wilde
}
