% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/na_in_row.R
\name{na_in_row}
\alias{na_in_row}
\title{NA in row}
\usage{
na_in_row(df, name = "na_in_row")
}
\arguments{
\item{df}{data frame}

\item{name}{name of new column}
}
\value{
tibble
}
\description{
Adds identifier based on whether rows contain NA in any column
}
\author{
Shona Wilde
}
