% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/negative_to_na.R
\name{negative_to_na}
\alias{negative_to_na}
\alias{zero_to_na}
\title{Functions to replace values with NA}
\usage{
negative_to_na(x)

zero_to_na(x)
}
\arguments{
\item{x}{Input vector}
}
\value{
Vector of same length as x.
}
\description{
Functions to replace values with NA
}
\author{
Shona Wilde
}
