% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalise.R
\name{normalise}
\alias{normalise}
\title{Function to normalise a numeric vector between 0 and 1}
\usage{
normalise(x, na.rm = TRUE)
}
\arguments{
\item{x}{Numeric vector}

\item{na.rm}{A logical indicating whether missing values should be removed. Default is TRUE.}
}
\value{
Numeric vector the same length as \code{x}.
}
\description{
Function to normalise a numeric vector between 0 and 1
}
\author{
Shona Wilde
}
