% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_unix_time.R
\name{parse_unix_time}
\alias{parse_unix_time}
\title{Function to parse unix time to a POSIXct date}
\usage{
parse_unix_time(x, tzone = "UTC", origin = "1970-01-01")
}
\arguments{
\item{x}{Unix time}

\item{tzone}{Character string containing the time zone to convert to}

\item{origin}{Origin of epoch. For unix time this is "1970-01-01"}
}
\value{
POSIXct date
}
\description{
Function to parse unix time to a POSIXct date
}
\author{
Shona Wilde
}
