% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paste_symbol.R
\name{paste_symbol}
\alias{paste_symbol}
\title{Function to paste symbols. Useful for labeling axis.}
\usage{
paste_symbol(x, symbol = "\%", position = "end", remove_space = T)
}
\arguments{
\item{x}{R object to be converted into a character vector}

\item{symbol}{Symbol. Default is %.}

\item{position}{Position to paste symbol relative to vector. Either "start" or "end".}
}
\value{
A character vector of the concatenated values.
}
\description{
Function to paste symbols. Useful for labeling axis.
}
\author{
Shona Wilde
}
