% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_colours.R
\name{plot_colours}
\alias{plot_colours}
\title{Plot colour palette
 
Function to visualize a colour palette}
\usage{
plot_colours(cols)
}
\arguments{
\item{cols}{Vector of colours}
}
\description{
Plot colour palette
 
Function to visualize a colour palette
}
\author{
Shona Wilde
}
