% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_strava_gps.R
\name{plot_strava_gps}
\alias{plot_strava_gps}
\title{Function to plot Strava GPS tracks from personal database}
\usage{
plot_strava_gps(activity_date = "2020-01-01")
}
\arguments{
\item{activity_date}{Date of activity/activities to plot}
}
\value{
\code{leaflet} map
}
\description{
Function to plot Strava GPS tracks from personal database
}
\author{
Shona Wilde
}
