% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_inf.R
\name{print_inf}
\alias{print_inf}
\title{Print infinate values}
\usage{
print_inf(x)
}
\arguments{
\item{x}{Input to print}
}
\value{
Tibble
}
\description{
Wrapper for \code{print(n = Inf)}
}
\author{
Shona Wilde
}
