% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_ceda_aq_file.R
\name{read_ceda_aq_file}
\alias{read_ceda_aq_file}
\title{Read CEDA Air Quality}
\usage{
read_ceda_aq_file(file, longer = FALSE, verbose = FALSE)
}
\arguments{
\item{file}{path to file.}

\item{longer}{Should the return be reshaped to be in a "tidy data" format? 
This operation is a bit slow.}

\item{verbose}{Should the function give messages on what file is being 
processed?}
}
\value{
Data frame/tibble
}
\description{
Reads and cleans CEDA AQ files
}
\author{
Shona Wilde
}
