% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_ceda_aq_file.R
\name{read_ceda_aq_files}
\alias{read_ceda_aq_files}
\title{Read CEDA AQ}
\usage{
read_ceda_aq_files(file)
}
\arguments{
\item{file}{path to file}
}
\description{
Reads and cleans CEDA AQ Files
}
\author{
Shona Wilde
}
