% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_ceda_ghg_file.R
\name{read_ceda_ghg_file}
\alias{read_ceda_ghg_file}
\title{Read CEDA Greenhouse Gas}
\usage{
read_ceda_ghg_file(file)
}
\arguments{
\item{file}{path to file}
}
\value{
data frame
}
\description{
Reads and cleans CEDA GHG files
}
\author{
Shona Wilde
}
