% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_ceda_met_file.R
\name{read_ceda_met_file}
\alias{read_ceda_met_file}
\title{Read CEDA Meteorology}
\usage{
read_ceda_met_file(file, site)
}
\arguments{
\item{file}{path to file}

\item{site}{"KM" or "LP"}
}
\value{
data frame
}
\description{
Reads and cleans CEDA met files
}
\author{
Shona Wilde
}
