% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_ceda_met_file.R
\name{read_ceda_met_files}
\alias{read_ceda_met_files}
\title{Read CEDA met}
\usage{
read_ceda_met_files(file)
}
\arguments{
\item{file}{path to file}
}
\value{
data frame
}
\description{
Reads and cleans CEDA met Files
}
\author{
Shona Wilde
}
