% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_core_faam_file.R
\name{read_core_faam_data}
\alias{read_core_faam_data}
\title{Read Core FAAM NetCDF}
\usage{
read_core_faam_data(core_file_list, verbose = T)
}
\arguments{
\item{core_file_list}{List containing core FAAM file names}
}
\value{
Tibble
}
\description{
Reads and cleans core FAAM files
}
\author{
Shona Wilde
}
