% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_core_faam_file.R
\name{read_core_faam_file}
\alias{read_core_faam_file}
\title{Read Core FAAM NetCDF}
\usage{
read_core_faam_file(file)
}
\arguments{
\item{file}{path to file}
}
\value{
data frame
}
\description{
Reads and cleans core FAAM files
}
\author{
Shona Wilde
}
