% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_faam_nitrates.R
\name{read_core_nitrates_data}
\alias{read_core_nitrates_data}
\title{Read  FAAM core nitrates file}
\usage{
read_core_nitrates_data(nitrates_file_list, verbose = TRUE)
}
\arguments{
\item{nitrates_file_list}{List containing core nitrates file names}
}
\value{
Tibble
}
\description{
Reads and cleans core nitrates FAAM file
}
\author{
Shona Wilde
}
