% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_core_van.R
\name{read_core_van}
\alias{read_core_van}
\title{Read core van data}
\usage{
read_core_van(file_list, tz = "UTC", force_utc = T, verbose = T)
}
\arguments{
\item{file_list}{files to read}

\item{tz}{a character string that specifies which time zone to parse the date with}

\item{force_utc}{should the date be converted to UTC?}

\item{verbose}{display message to user?}
}
\value{
Tibble.
}
\description{
Reads core data from the WACL
}
\author{
Shona Wilde
}
