% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_ecmwf.R
\name{read_ecmwf}
\alias{read_ecmwf}
\title{Read ECMWF}
\usage{
read_ecmwf(file_list, verbose = T)
}
\arguments{
\item{file_list}{Files to read}

\item{verbose}{Should a message to the user be displayed?}
}
\value{
Tibble
}
\description{
Reads and cleans data from the Copernicus Climate Data Store (CDS)
}
\author{
Shona Wilde
}
