% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_faam_flight_summary.R
\name{read_faam_flight_summary}
\alias{read_faam_flight_summary}
\title{Read FAAM flight summary}
\usage{
read_faam_flight_summary(file, verbose)
}
\arguments{
\item{x}{file}
}
\value{
df
}
\description{
Reads and cleans flight summary from aircraft
}
\author{
Stuart Grange
}
