% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_fgga_data.R
\name{read_fgga_data}
\alias{read_fgga_data}
\title{Read NASA AMES FGGA file from FAAM aircraft}
\usage{
read_fgga_data(fgga_file_list, verbose = T)
}
\arguments{
\item{core_file_list}{List containing FGGA files}
}
\value{
Tibble
}
\description{
Reads and cleans FGGA files from FAAM aircraft
}
\author{
Shona Wilde
}
