% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_gc_werks.R
\name{read_gc_werks}
\alias{read_gc_werks}
\title{Read GCwerks}
\usage{
read_gc_werks(file_list, verbose = T)
}
\arguments{
\item{file_list}{List containing file names}

\item{verbose}{Should a message to the user be displayed?}
}
\value{
Tibble
}
\description{
Reads and cleans raw data files from the GCwerks program
}
\author{
Shona Wilde
}
