% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_genesis.R
\name{read_genesis}
\alias{read_genesis}
\title{Read Genesis}
\usage{
read_genesis(file_list, verbose = TRUE)
}
\arguments{
\item{file_list}{Files to read}

\item{verbose}{Should a message to the user be displayed?}
}
\value{
Tibble
}
\description{
Reads and cleans process air survey emissions calculation data sheets
}
\author{
Shona Wilde
}
