% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_icad_data.R
\name{read_icad_data}
\alias{read_icad_data}
\title{Read data from the ICAD instrument}
\usage{
read_icad_data(file_list, pad = T, verbose = T)
}
\arguments{
\item{file_list}{List of files to read}

\item{pad}{Should the time series be padded? If TRUE the time series is interpolated to give an observation every second}

\item{verbose}{Display message to user?}
}
\value{
Tibble
}
\description{
Read data from the ICAD instrument
}
\author{
Shona Wilde
}
