% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_icl_str_file.R
\name{read_icl_str_file}
\alias{read_icl_str_file}
\title{Calculate Read raw ICL file}
\usage{
read_icl_str_file(file)
}
\arguments{
\item{file}{file}
}
\value{
tibble
}
\description{
Reads raw data str data file from Aerodyne ICL
}
\author{
Shona Wilde
}
