% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_markes_data.R
\name{read_markes_data}
\alias{read_markes_data}
\title{Read Markes data}
\usage{
read_markes_data(file_list, verbose = T)
}
\arguments{
\item{file_list}{List containing file names}

\item{verbose}{Should a message to the user be displayed?}
}
\value{
Tibble
}
\description{
Reads and cleans raw data files from Markes pre-concentration units
}
\author{
Shona Wilde
}
