% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_mass_flux_output.R
\name{read_mass_flux_output}
\alias{read_mass_flux_output}
\title{Read mass flux output}
\usage{
read_mass_flux_output(file_list)
}
\arguments{
\item{file_list}{List containing file names}
}
\value{
Tibble
}
\description{
Reads and cleans data files from the MatLab mass flux script
}
\author{
Shona Wilde
}
