% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_merge_files.R
\name{read_merge_file}
\alias{read_merge_file}
\title{Read FAAM merge files}
\usage{
read_merge_file(file_list, verbose = T)
}
\arguments{
\item{verbose}{should messgaes print in console}
}
\value{
tibble
}
\description{
Reads and cleans FAAM merge files
}
\details{
#' @param file_list list of merge files to read
}
\author{
Shona Wilde
}
