% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_syft_data.R
\name{read_syft_data}
\alias{read_syft_data}
\title{Read SYFT data}
\usage{
read_syft_data(file)
}
\arguments{
\item{file}{file}
}
\value{
tibble
}
\description{
Reads worked up SYFT data
}
\author{
Shona Wilde
}
