% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_td_data.R
\name{read_td_data}
\alias{read_td_data}
\title{Read TD data}
\usage{
read_td_data(file_list, verbose = T)
}
\arguments{
\item{file_list}{List containing file names}

\item{verbose}{Should a message to the user be displayed?}
}
\value{
Tibble
}
\description{
Reads and cleans raw data files from the WACL thermal desorption (TD) units
}
\author{
Shona Wilde
}
