% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_unnamed_td_data.R
\name{read_unnamed_td_data}
\alias{read_unnamed_td_data}
\title{Read unnamed TD data}
\usage{
read_unnamed_td_data(file_list, verbose = T)
}
\arguments{
\item{file_list}{List of files to read}

\item{verbose}{Display message to user?}
}
\value{
Tibble
}
\description{
Function to read data from the WACL thermal desorption units with unnamed columns.
}
\author{
Shona Wilde
}
