% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_was_log.R
\name{read_was_log}
\alias{read_was_log}
\title{Read WAS Log}
\usage{
read_was_log(file)
}
\arguments{
\item{file}{file to read of type ".WAS"}
}
\value{
df
}
\description{
Reads and cleans WAS log file from FAAM aircraft
}
\author{
Stuart Grange / Shona Wilde
}
