% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_outliers.R
\name{remove_outliers}
\alias{remove_outliers}
\title{Function to remove outliers in a data frame.}
\usage{
remove_outliers(df, x, coef = 10)
}
\arguments{
\item{df}{Data frame}

\item{x}{Numeric vector within \code{df}}

\item{coef}{A coefficient which determines the classification of an outlier. Use 0 for no outliers to be identified.}
}
\description{
Function to remove outliers in a data frame.
}
\examples{
\dontrun{

df \%>\% group_by(variable) \%>\% remove_outliers(value)

}

}
\author{
Shona Wilde
}
