% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replace_na_unique.R
\name{replace_na_unique}
\alias{replace_na_unique}
\title{Function to replace NAs with unique numbers}
\usage{
replace_na_unique(x, min_number = 1)
}
\arguments{
\item{x}{Numeric vector.}

\item{min_number}{Minimum number to start replacement from.}
}
\value{
Numeric vector of same length as \code{x}.
}
\description{
Function to replace NAs with unique numbers
}
