% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rma.R, R/rma_model.R
\name{rma}
\alias{rma}
\title{Model2 regression}
\usage{
rma(df, formula)

rma(df, formula)
}
\arguments{
\item{df}{Data frame containing model variables}

\item{formula}{A formula specifying the bivariate model, as in \code{\link[stats]{lm} and \code{\link[lmodel2]{lmodel2}}}}
}
\value{
Regression output

Regression output
}
\description{
Generic wrapper for the \code{lmodel2} function. Should be used when there is an error associated with both the x and y variables in the regression

Generic wrapper for the \code{lmodel2} function. Should be used when there is an error associated with both the x and y variables in the regression
}
\author{
Shona Wilde

Shona Wilde
}
