% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model2_regression.R
\name{rma_model}
\alias{rma_model}
\title{Model2 regression}
\usage{
rma_model(df, x = "x", y = "y")
}
\arguments{
\item{df}{Data frame containing model variables}

\item{x}{x-component for model}

\item{y}{x-component for model}
}
\value{
model output
}
\description{
Generic wrapper for the \code{lmodel2} function
}
\author{
Shona Wilde
}
