% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sethd.R
\name{sethd}
\alias{sethd}
\title{Set home directory}
\usage{
sethd()
}
\value{
\code{sethd} returns the current directory before the change as in \code{setwd}.
}
\description{
Set home directory based on the current operating system
}
\author{
Shona Wilde
}
