% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sfc_to_columns.R
\name{sfc_to_columns}
\alias{sfc_to_columns}
\title{Function to extract geometry coordinates as columns}
\usage{
sfc_to_columns(data, names = c("longitude", "latitude"), drop_geometry = FALSE)
}
\arguments{
\item{data}{Data with class \code{sf} or \code{sfc}}

\item{names}{Names of new columns}

\item{drop_geometry}{Should the `geometry` column be dropped?}
}
\value{
Either a `tibble` or `sf` object
}
\description{
Function to extract geometry coordinates as columns
}
\author{
Shona Wilde
}
