% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shonarrr-package.R
\docType{package}
\name{shonarrr-package}
\alias{shonarrr}
\alias{shonarrr-package}
\title{shonarrr: shonarrr}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/shonawilde/shonarrr}
}

}
\keyword{internal}
