% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slice_by_indicies.R
\name{slice_by_indices}
\alias{slice_by_indices}
\title{Slice by a range of indexes}
\usage{
slice_by_indices(df, index_start, index_end)
}
\arguments{
\item{df}{Input data frame or tibble}

\item{index_start}{Start row value}

\item{index_end}{End row value}
}
\value{
Tibble
}
\description{
Slice by a range of indexes
}
\author{
Shona Wilde
}
