% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slice_min_max.R
\name{slice_min_max}
\alias{slice_min_max}
\title{Slice min and max rows}
\usage{
slice_min_max(df, order_by = value, n = 1)
}
\arguments{
\item{df}{data}

\item{order_by}{variable to order by}

\item{n}{number of rows to select, default is 1}
}
\value{
tibble
}
\description{
Select rows of highest and lowest values based on variables
}
\author{
Shona Wilde
}
