% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st_distinct.R
\name{st_distinct}
\alias{st_distinct}
\title{Function to find distinct rows from a spatial object}
\usage{
st_distinct(obj, ..., .keep_all = FALSE)
}
\arguments{
\item{obj}{object of class \code{sf} or \code{sfc}}

\item{...}{Optional variables to use when determining uniqueness.}

\item{.keep_all}{If \code{TRUE}, keep all variables in \code{obj}.}
}
\value{
df
}
\description{
Function to find distinct rows from a spatial object
}
\author{
Shona Wilde
}
