% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st_from_df.R
\name{st_from_df}
\alias{st_from_df}
\title{Function to extract geometry coordinates as columns}
\usage{
st_from_df(df, latitude = "latitude", longitude = "longitude", crs = 4326)
}
\arguments{
\item{df}{Data frame to be converted to spatial data frame - currently only works for `POINTS`}

\item{latitude}{Latitude varaible in df}

\item{longitude}{Longitude varaible in df}

\item{crs}{Coordinate reference system of df}
}
\value{
`sf`object
}
\description{
Function to extract geometry coordinates as columns
}
\author{
Shona Wilde
}
