% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str_extract_round_brackets.R
\name{str_extract_round_brackets}
\alias{str_extract_round_brackets}
\alias{str_extract_square_brackets}
\alias{str_rm_unicode}
\title{Functions for string manipulation}
\usage{
str_extract_round_brackets(x)

str_extract_square_brackets(x)

str_rm_unicode(x)
}
\arguments{
\item{x}{Input string.}
}
\value{
Output string.
}
\description{
Functions for string manipulation
}
\author{
Shona Wilde.
}
