% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str_first_to_last.R
\name{str_first_to_last}
\alias{str_first_to_last}
\title{Extract first word of string and move to end}
\usage{
str_first_to_last(x)
}
\arguments{
\item{x}{Input string}
}
\value{
Output string
}
\description{
Extract first word of string and move to end
}
\author{
Shona Wilde
}
