% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str_split_extract_element.R
\name{str_split_extract_element}
\alias{str_split_extract_element}
\title{Function to extract a specific element from \code{str_split}}
\usage{
str_split_extract_element(x, pattern, n = Inf, element = 1, as_character = T)
}
\arguments{
\item{x}{Input string}

\item{pattern}{Pattern to split string on}

\item{n}{Number of pieces to return. Default (Inf) uses all possible split positions}

\item{element}{Element of split string to return}

\item{as_character}{Should the string be returned as a character?}
}
\description{
Function to extract a specific element from \code{str_split}
}
\author{
Shona Wilde
}
