% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str_tidy.R
\name{str_tidy}
\alias{str_tidy}
\title{Converts strings to a tidy format}
\usage{
str_tidy(x)
}
\arguments{
\item{x}{Input string}
}
\value{
Cleaned string
}
\description{
Converts strings to a tidy format
}
\author{
Shona Wilde
}
