% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sunrise_sunset.R
\name{sunrise_sunset}
\alias{sunrise_sunset}
\title{Function to calculate sunrise and sunset times at a particular location}
\usage{
sunrise_sunset(date, latitude, longitude)
}
\arguments{
\item{latitude}{Latitude of location}

\item{longitude}{Longitude of location}
}
\value{
Tibble.
}
\description{
Function to calculate sunrise and sunset times at a particular location
}
\examples{
# append to data frame
df \%>\% 
  mutate(sunrise_sunset(date, 54.20092, -0.81852))

}
\seealso{
\code{\link{sunriset}}, \code{\link{gissr::calculate_sunrise}}
}
\author{
Shona Wilde
}
