% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_shona_less_minimal.R
\name{theme_shona_less_minimal}
\alias{theme_shona_less_minimal}
\title{Theme Less Minimal Big}
\usage{
theme_shona_less_minimal(legend_position = "right")
}
\description{
theme_minimal but bigger and less minimal
}
\author{
Shona Wilde
}
