% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_shona_void.R
\name{theme_shona_void}
\alias{theme_shona_void}
\title{Theme Shona Void}
\usage{
theme_shona_void(legend_position = "top")
}
\description{
Void theme designed for use with maps
}
\author{
Shona Wilde
}
