% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_shona_white_big.R
\name{theme_shona_white_big}
\alias{theme_shona_white_big}
\title{Theme Minimal Big}
\usage{
theme_shona_white_big(legend_position = "right")
}
\description{
theme_minimal but bigger
}
\author{
Shona Wilde
}
