% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_shona_white_journal.R
\name{theme_shona_white_journal}
\alias{theme_shona_white_journal}
\title{A smart, spacy theme for a ggplot}
\usage{
theme_shona_white_journal(
  legend_position = "right",
  grid_type = "dashed",
  strip = T,
  text_size = 9
)
}
\arguments{
\item{legend_position}{Location of legend - choices are "top", "bottom", "right", or  "left".}

\item{grid_type}{Line type of \code{panel.grid}. An integer (0:8), a name (blank, solid, dashed, dotted, dotdash, longdash, twodash), 
or a string with an even number (up to eight) of hexadecimal digits which give the lengths in consecutive positions in the string.
If grid_type = "none", no grid will be displayed}

\item{text_size}{Size of plot text.}
}
\description{
Designed to produce publication-quality figures with appropriate text sizes
}
\author{
Shona Wilde
}
