% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_barnes_output.R
\name{tidy_barnes_output}
\alias{tidy_barnes_output}
\title{Tidy banres output}
\usage{
tidy_barnes_output(file)
}
\arguments{
\item{file}{File to read}
}
\value{
Tibble
}
\description{
Reads and re-formats MatLab data from Barnes interpolation
}
\author{
Shona Wilde
}
