% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_gam_output.R
\name{tidy_gam_output}
\alias{tidy_gam_output}
\title{Function to summarise the output from \code{\link[mgcv]{gam}} in a tidy format}
\usage{
tidy_gam_output(model)
}
\arguments{
\item{model}{Model object.}
}
\value{
Tibble containing model results.
}
\description{
Function to summarise the output from \code{\link[mgcv]{gam}} in a tidy format
}
\author{
Shona Wilde
}
