% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_lm_output.R
\name{tidy_lm_output}
\alias{tidy_lm_output}
\title{Function to summarise the output from \code{\link[stats]{lm}} in a tidy format}
\usage{
tidy_lm_output(model)
}
\arguments{
\item{model}{Model object to summarise.}
}
\value{
Tibble containing model results.
}
\description{
Function to summarise the output from \code{\link[stats]{lm}} in a tidy format
}
\author{
Shona Wilde
}
