% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_model_summary.R
\name{tidy_model_summary}
\alias{tidy_model_summary}
\title{Tidy model summary}
\usage{
tidy_model_summary(model)
}
\arguments{
\item{model}{Model for summarising}
}
\value{
Tibble
}
\description{
Combines useful model summary functions \code{broom::glance} and \code{broom::tidy} to produce a single tidy summary 
for linear models
}
\author{
Shona Wilde
}
