% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_rma_output.R
\name{tidy_rma_output}
\alias{tidy_rma_output}
\title{Tidy RMA output}
\usage{
tidy_rma_output(model)
}
\arguments{
\item{model}{Model for summarising}
}
\value{
Tibble
}
\description{
Combines useful model summary functions \code{broom::glance} and \code{broom::tidy} to produce a single tidy summary 
from a \code{\link[lmodel2]{lmodel2}} object.
}
\author{
Shona Wilde
}
