% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_rq_output.R
\name{tidy_rq_output}
\alias{tidy_rq_output}
\title{Function to summarise the output from \code{\link[quantreg]{rq}} in a tidy format}
\usage{
tidy_rq_output(model, se = NULL)
}
\arguments{
\item{model}{Model object to summarise.}

\item{se}{Method to use for calculating standard errors. One of NULL, "nid", "rank", "iid", "ker", or "boot".}
}
\value{
Tibble containing model results.
}
\description{
Function to summarise the output from \code{\link[quantreg]{rq}} in a tidy format
}
\author{
Shona Wilde
}
