% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_since_uk_lockdown.R
\name{time_since_uk_lockdown}
\alias{time_since_uk_lockdown}
\title{Function to calculate time since UK lockdown}
\usage{
time_since_uk_lockdown(format = "days")
}
\arguments{
\item{format}{Type of return. Can be "seconds", "hms", "duration", "period" or "days" (default).}
}
\value{
Vector
}
\description{
Function to calculate time since UK lockdown
}
\author{
Shona Wilde
}
