% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_alpha.R
\name{calc_alpha}
\alias{calc_alpha}
\title{Calculate alpha}
\usage{
calc_alpha(wind_direction, heading)
}
\arguments{
\item{wind_direction}{Wind direction in degrees}

\item{heading}{Aircraft heading in degrees}
}
\value{
value
}
\description{
Calculate alpha which is the angle between aircraft track and direction normal to wind direction
}
\author{
Shona Wilde
}
