% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_gps.R
\name{conv_gps}
\alias{conv_gps}
\title{Convert GPS}
\usage{
conv_gps(coord, direction)
}
\arguments{
\item{coord}{latitude or longitude column to be converted}

\item{direction}{column containing directional component of coordinates (N, S, E, W)}
}
\value{
Numeric of converted coordinate
}
\description{
Convert GPS data from WACL van to decimal degrees
}
\author{
Shona Wilde
}
