% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_ceda_aq_file.R
\name{read_ceda_aq_file}
\alias{read_ceda_aq_file}
\title{Read CEDA Air Quality}
\usage{
read_ceda_aq_file(file)
}
\arguments{
\item{file}{path to file}
}
\value{
data frame
}
\description{
Reads and cleans CEDA AQ files
}
\author{
Shona Wilde
}
