% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_core_van.R
\name{read_core_van}
\alias{read_core_van}
\title{Calculate Reads core van}
\usage{
read_core_van(file)
}
\arguments{
\item{file}{file}
}
\value{
tibble
}
\description{
Reads core data fromm WACL van
}
\author{
Shona Wilde
}
