% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_faam_flight_summary.R
\name{read_faam_flight_summary}
\alias{read_faam_flight_summary}
\title{Read FAAM flight summary}
\usage{
read_faam_flight_summary(flight_sum_list, create_end_time = 10,
  verbose = T)
}
\arguments{
\item{flight_sum_list}{List of flight summaries}

\item{create_end_time}{Time in minutes to manually create an end time for range events
if these are missing in the flight summary. End times are required for \code{flight_range_subset}.
Defualt is 10 mins as this is a typical length of an aircraft run.}
}
\value{
df
}
\description{
Reads and cleans flight summary from aircraft
}
\author{
Stuart Grange / Shona Wilde
}
