% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rma_model.R
\name{rma}
\alias{rma}
\title{Model2 regression}
\usage{
rma(formula, data, ...)
}
\arguments{
\item{formula}{A formula specifying the bivariate model, as in \code{\link[stats]{lm} and \code{\link[lmodel2]{lmodel2}}}}

\item{data}{Input data containing model variables}

\item{...}{Additional arguments to be passed to \code{lmodel2}}
}
\value{
Model object
}
\description{
Generic wrapper for the \code{lmodel2} function. Should be used when there is an error associated with both the x and y variables in the regression
}
\author{
Shona Wilde
}
