% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str_extract_round_brackets.R
\name{str_extract_round_brackets}
\alias{str_extract_round_brackets}
\title{Extract text within round brackets}
\usage{
str_extract_round_brackets(x)
}
\arguments{
\item{x}{string}
}
\value{
string
}
\description{
Extracts texts contained within round brackets in character strings
}
\author{
Shona Wilde
}
