% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_barnes_output.R
\name{tidy_barnes_output}
\alias{tidy_barnes_output}
\title{Tidy banres output}
\usage{
tidy_barnes_output(file_list, verbose = T)
}
\arguments{
\item{file_list}{Files to read}

\item{verbose}{Should a message to the user be displayed?}
}
\value{
Tibble
}
\description{
Reads and re-formats MatLab data from Barnes interpolation
}
\author{
Shona Wilde
}
